+++ src/DocMeasure.js
@@ -95,8 +95,10 @@ class DocMeasure {
 			node._width = node._minWidth = node._maxWidth = node.cover.width;
 			node._height = node._minHeight = node._maxHeight = node.cover.height;
 		} else {
-			node._width = node._minWidth = node._maxWidth = node.width || dimensions.width;
-			node._height = node.height || (dimensions.height * node._width / dimensions.width);
+			let ratio = dimensions.width / dimensions.height;
+
+			node._width = node._minWidth = node._maxWidth = node.width || (node.height ? (node.height * ratio) : dimensions.width);
+			node._height = node.height || (node.width ? node.width / ratio : dimensions.height);
 
 			if (isNumber(node.maxWidth) && node.maxWidth < node._width) {
 				node._width = node._minWidth = node._maxWidth = node.maxWidth;