+++ src/DocMeasure.js
@@ -248,7 +248,7 @@ class DocMeasure {
 		return this.textInlines.sizeOfText('9. ', this.styleStack);
 	}
 
-	buildUnorderedMarker(styleStack, gapSize, type) {
+	buildUnorderedMarker(item, styleStack, gapSize, type) {
 		function buildDisc(gapSize, color) {
 			// TODO: ascender-based calculations
 			let radius = gapSize.fontSize / 6;
@@ -295,7 +295,7 @@ class DocMeasure {
 		}
 
 		let marker;
-		let color = styleStack.getProperty('markerColor') || styleStack.getProperty('color') || 'black';
+		let color = StyleContextStack.getStyleProperty(item, styleStack, 'markerColor', undefined) || styleStack.getProperty('color') || 'black';
 
 		switch (type) {
 			case 'circle':
@@ -322,7 +322,7 @@ class DocMeasure {
 		return marker;
 	}
 
-	buildOrderedMarker(counter, styleStack, type, separator) {
+	buildOrderedMarker(item, counter, styleStack, type, separator) {
 		function prepareAlpha(counter) {
 			function toAlpha(num) {
 				return (num >= 26 ? toAlpha((num / 26 >> 0) - 1) : '') + 'abcdefghijklmnopqrstuvwxyz'[num % 26 >> 0];
@@ -402,11 +402,11 @@ class DocMeasure {
 			}
 		}
 
-		let textArray = { text: counterText };
-		let markerColor = styleStack.getProperty('markerColor');
-		if (markerColor) {
-			textArray.color = markerColor;
-		}
+		let markerColor = StyleContextStack.getStyleProperty(item, styleStack, 'markerColor', undefined) || styleStack.getProperty('color') || 'black';
+		let textArray = {
+			text: counterText,
+			color: markerColor
+		};
 
 		return { _inlines: this.textInlines.buildInlines(textArray, styleStack).items };
 	}
@@ -423,7 +423,7 @@ class DocMeasure {
 			let item = items[i] = this.measureNode(items[i]);
 
 			if (!item.ol && !item.ul) {
-				item.listMarker = this.buildUnorderedMarker(style, node._gapSize, item.listType || node.type);
+				item.listMarker = this.buildUnorderedMarker(item, style, node._gapSize, item.listType || node.type);
 			}
 
 			node._minWidth = Math.max(node._minWidth, items[i]._minWidth + node._gapSize.width);
@@ -452,7 +452,7 @@ class DocMeasure {
 
 			if (!item.ol && !item.ul) {
 				let counterValue = isNumber(item.counter) ? item.counter : counter;
-				item.listMarker = this.buildOrderedMarker(counterValue, style, item.listType || node.type, node.separator);
+				item.listMarker = this.buildOrderedMarker(item, counterValue, style, item.listType || node.type, node.separator);
 				if (item.listMarker._inlines) {
 					node._gapSize.width = Math.max(node._gapSize.width, item.listMarker._inlines[0].width);
 				}